[COMM]: 修改Readn和Writen的返回类型和参数类型

+++ src/incl/redo.h
@@ -7,8 +7,8 @@
 #define OPEN_FLAGS          (O_CREAT|O_WRONLY|O_APPEND) /* 文件标志 */
 #define DIR_MODE            (0777)  /* 目录权值 */
 
-int Readn(int fd, void *buff, int n);
-int Writen(int fd, const void *buff, int n);
+ssize_t Readn(int fd, void *buff, size_t n);
+ssize_t Writen(int fd, const void *buff, size_t n);
 int Open(const char *fpath, int flags, mode_t mode);
 #define CLOSE(fd)  { if(fd > 0) { close(fd), fd = -1; }}
 #define FCLOSE(fp) {fclose(fp), fp = NULL;}
